+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ Version 2021.X (2021-XX-XX)
 * Allow 3D transforms on arrays with >3 dimensions (batch transform)
 * Support for transforms on a given list of axes, instead of
   only the first ndim axes. Unavailable for R2C.
+* Direct Cosine Transform (DCT) of type 2, 3 and 4.
 * OpenCL: test for half and double-precision support
 * OpenCL: relax accuracy requirements in unit tests
 * Fix shape test for out-of-place R2C transforms